update/delete working

+++ src/components/Admin/Users/index.js
@@ -7,6 +7,9 @@ class Users extends React.Component {
 
     constructor(props) {
         super(props);
+        this.state = {
+            editUser: null
+        }
     }
 
     componentDidMount() {
@@ -14,7 +17,19 @@ class Users extends React.Component {
     }
 
     deleteHandler(id) {
-        this.props.deleteUser(this.props.token, id);
+        const deleteConfirm = window.confirm("Are you sure you want to delete this user?");
+        if(deleteConfirm) {
+            this.props.deleteUser(this.props.token, id);
+        }
+    }
+
+    editHandler(user) {
+        this.setState({editUser: user});
+    }
+
+    updateHandler(token, user) {
+        this.props.updateUser(this.props.token, user);
+        this.setState({editUser: null});
     }
 
     render() {
@@ -35,17 +50,23 @@ class Users extends React.Component {
                                 <UserListItem 
                                     key={idx}
                                     user={user}
-                                    editHandler={()=>{console.log("EDIT")}}
+                                    editHandler={this.editHandler.bind(this)}
                                     deleteHandler={this.deleteHandler.bind(this)}
                                 />
                             );
                         })
                     }
                 </ul> 
-                <UserForm 
-                    getUsers={this.props.getUsers} 
-                    addUser={this.props.addUser} 
+                <UserForm
+                    title="Add User" 
+                    submitHandler={this.props.addUser} 
                     token={this.props.token} 
+                />
+                <UserForm 
+                    title="Edit User"
+                    submitHandler={this.updateHandler.bind(this)}
+                    token={this.props.token}
+                    user={this.state.editUser}
                 />
             </div>
         );
